+++ litmus-portal/graphql-server/pkg/authorization/validate.go
@@ -1,4 +1,4 @@
-package validate
+package authorization
 
 import (
 	"context"
@@ -9,19 +9,18 @@ import (
 	"go.mongodb.org/mongo-driver/bson"
 
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/graph/model"
-	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/authorization"
 )
 
-// ValidateRole :Validates the role of a user in a given project
+// ValidateRole Validates the role of a user in a given project
 func ValidateRole(ctx context.Context, projectID string, requiredRoles []model.MemberRole, invitation string) error {
-	claims := ctx.Value(authorization.UserClaim).(jwt.MapClaims)
+	claims := ctx.Value(UserClaim).(jwt.MapClaims)
 	uid := claims["uid"].(string)
 
 	filter := bson.D{{"members", bson.D{{"$elemMatch", bson.D{{"user_id", uid}, {"role", bson.D{{"$in", requiredRoles}}}, {"invitation", invitation}}}}}, {"_id", projectID}}
 	_, err := dbOperationsProject.GetProject(ctx, filter)
 
 	if err != nil {
-		return errors.New("Permission Denied")
+		return errors.New("permission denied")
 	}
 
 	return nil
